refactor: withdrawal
