+++ src/main/java/com/spacestar/back/auth/service/MemberService.java
@@ -1,0 +1,8 @@
+package com.spacestar.back.auth.service;
+
+public interface MemberService {
+
+    void withdrawal(String uuid);
+
+    void withdrawalForce(String uuid);
+}
